third_party/WebKit/LayoutTests/external/wpt/css/css-shapes-1/shape-outside/values/shape-image-threshold-003-expected.txt

This is a testharness.js-based test.
PASS shape-image-threshold can be assigned 'inherit' value
FAIL shape-image-threshold is not inherited and defaults to 0assert_equals: expected (object) null but got (string) ""
PASS shape-margin is not inherited
Harness: the test ran to completion.
